fix: import and export media and commands defer reply

+++ src/rankflix-console/src/main/java/pt/graca/discord/command/media/ConfirmMediaButtonListener.java
@@ -1,6 +1,7 @@
 package pt.graca.discord.command.media;
 
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.interactions.components.selections.EntitySelectMenu;
@@ -24,35 +25,48 @@ public class ConfirmMediaButtonListener extends ListenerAdapter {
 
             if (componentId.startsWith("confirm-media")) {
                 event.deferEdit().queue();
-                int mediaTmdbId = Integer.parseInt(componentId.split(":")[1]);
-                service.addMedia(mediaTmdbId);
+
+                if (isNotAdmin(event)) {
+                    throw new IllegalAccessException("You must be an administrator to use this command");
+                }
+
+                String mediaId = componentId.split(":")[1];
+                service.addMedia(mediaId);
 
                 // keep links, delete confirm buttons
                 var newComponents = event.getMessage().getActionRows().getFirst();
                 event.getHook().editOriginalComponents(newComponents).queue();
 
-                sendWatchersSelectorEmbed(event, mediaTmdbId);
+                sendWatchersSelectorEmbed(event, mediaId);
             } else if (componentId.equals("ignore-media")) {
                 event.deferEdit().queue();
+
+                if (isNotAdmin(event)) {
+                    throw new IllegalAccessException("You must be an administrator to use this command");
+                }
+
                 event.getHook().deleteMessageById(event.getMessageId()).queue();
             }
         } catch (Exception e) {
             e.printStackTrace();
-            event.replyEmbeds(new EmbedBuilder()
-                            .setDescription(e.getMessage())
-                            .setColor(Color.RED)
-                            .build()
-                    )
-                    .setEphemeral(true)
-                    .queue();
+
+            event.getHook().sendMessageEmbeds(new EmbedBuilder()
+                    .setTitle("Error")
+                    .setDescription(e.getMessage() != null ? e.getMessage() : "Internal error")
+                    .setColor(Color.RED)
+                    .build()
+            ).setEphemeral(true).queue();
         }
     }
 
-    private void sendWatchersSelectorEmbed(ButtonInteractionEvent event, int mediaTmdbId) {
+    private boolean isNotAdmin(ButtonInteractionEvent event) {
+        return event.getMember() == null || !event.getMember().hasPermission(Permission.ADMINISTRATOR);
+    }
+
+    private void sendWatchersSelectorEmbed(ButtonInteractionEvent event, String mediaId) {
         EntitySelectMenu selectMenu = EntitySelectMenu.create(
-                        "add-media-users-select:" + mediaTmdbId,
-                        EntitySelectMenu.SelectTarget.USER
-                )
+                        "add-media-users-select:" + mediaId,
+                        EntitySelectMenu.SelectTarget.USER)
                 .setMaxValues(25) // Discord limit
                 .setPlaceholder("Pick users who watched it")
                 .build();
